pmem-vgm: Fix issue in device grouping
We were expecting that pvdisplay returns error if it run on the existing physical
volume which is not part of any volume group, but our assumption was wrong.

To fix this case now we run "pvs -o vg_name" command on every physical volume it
returns it's exact volume group name if any other wise null string.

Signed-off-by: Amarnath Valluri <amarnath.valluri@intel.com>