test: bring up Clear Linux under QEMU
This automates the process of installing Clear Linux on different
virtual machine instances and then bringing up a Kubernetes cluster.
Except for the initial networking setup everything runs as the current
user. The runqemu-ifup/down scripts get started as normal user, but
then depend on sudo.

Docker is used despite not being the default in Clear Linux because it
allowed running the Kubernetes cluster without a network
plugin. Installing a network plugin slows down the cluster boot time
without adding any (for the purpose of the test cluster) relevant
advantage.

.gitignore (was missing) and .dockerignore are getting added. The
latter is important because otherwise the entire current directory,
including the new and large _work, are sent to the Docker daemon.

The example deployment file is getting updated so that it can work
directly on the test cluster with a local Docker registry, the one
which is also the default for "make push-images".