README.md: QEMU cluster and E2E testing, cluster limitations
Besides adding a completely new section for developers at the end, it
also clarifies a few things in the existing sections:
- Docker registry dependency
- remove "Verify socket" section (nothing coming before it
  shows how to start the CSI driver, so there isn't anything
  to verify)
- remove the "Test plugin" section (the usage of the
  installed plugin was already explained in the two preceeding
  sections)
- csi-pmem not guaranteed to work on cluster with multiple
  nodes that have NVDIMM (*)

(*) A volume gets created when a pod starts using the PV, on the node
where the pod is about to run, and deleted when the PV gets
deleted. If the pod restarts or a second pod starts using the PV, it
is not guaranteed that the pod will run on the node where the volume
was created initially.