update Kubernetes, remove glog
This is in preparation for using Kubernetes E2E storage
tests. Kubernetes 1.13 replaced glog with a fork, out of concerns
regarding the upstreams maintenance. It makes sense to follow that
change and stop using glog completely.

The out-of-tree version of the "testsuites" package has to be used
because two PRs are still under review:
https://github.com/kubernetes/kubernetes/pull/70862
https://github.com/kubernetes/kubernetes/pull/70992

Kubernetes 1.13 could be used instead, but then more code would get
pulled in (PR 70862) or the API would be different (PR 70992). Either
way, the implementation of k8s.io/kubernetes/pkg/util/mount
changes, which affects the pmem-csi-driver.