Offload GetCapacity() from RegistryServer to Node ControllerServer
Only nodes can provide reliable value about 'available capacity' on a node at a
given point of time. Book-keeping this at registry server on master side is not
reliable and harder to keep in sync/updated.

So more reliable approach is that Master Controller gets the node capacity from
Node Controller on need basis. CSI ControllerServer interface already have
GetCapacity() method, so this change makes use of this API.

Now we can clearly define the use of RegistryServer interface that its only
responsibility is to maintain a registry of node controller server's endpoints.

So, with this change, Node controller servers implements
ControllerServer.GetCapacity() and Master controller uses this method to
determine available capacity by a node.

Updated the sequence diagram to show this additional step.