Deployment: add manifest for direct-nvdimm mode
New yaml file specifies use of driver in direct mode.
Two pre-stages are skipped, devicemode arg is added.
We change existing pmem-csi.yaml to be symlink to
original deployment manifest that we rename to be lvm specific,
to maintain existing use cases pointing to it
and reduce parallel documenting needs.