deployment: clarify Kubernetes versioning
The documentation and the naming looked like `deploy/kubernetes` was
to be used unless stated otherwise, but in reality that was for old
Kubernetes versions that are probably not the right choice anymore.

Now the Kubernetes release number is explicit in each deploy
directory. When there is no change needed (as for 1.13), a symlink
can be used.