test/e2e: use already deployed pmem-csi driver
Deploying anew for each test namespace led to an unusable driver
because the TLS secrets were missing. Long-term the goal upstream is
to move away from per-test deployments and instead test the more
realistic scenario of keeping a driver running across tests, so here
we do the same.

An additional advantage is that tests execute faster.

The Gingko test log output still includes the logs of the pmem-csi
driver, by listing everything that runs in the default namespace. This
works well as long as only one test runs at a time, but will be less
useful for parallel execution of tests.