Alignment and size pre-increase changes
Caused by how libndctl creates namespaces, we have to adjust
our creation request.
libndctl v63 creates 1GB smaller size than we request,
if alignment is set to 1GB, and fails to create if request
is smaller than 2 GB. We have to pass sanity testing where small
volumes are created as well.
In devicemode:LVM2, we want to allocate as much as we can, but
we don't care about sizes.
In devicemode:direct, we want the size to be not less what we want.
In both cases, we want 1GB alignment.