e2e sanity: fix deployment of unified driver
Since commit 8054230bbfedadb282336dcfabb03e7c776c17ae, the "normal"
pmem-csi driver is already installed when testing runs, while the
"unified" driver still gets installed once per test. For that to work,
the service and app definitions in the pmem-unified.yaml must not
conflict with the ones for the normal driver.

In addition, the temporary target and staging directories in /var/tmp
must also be visible inside the container.