pmd-ndctl.go: Add start-time check for writable /sys
In Devicemode_direct (alias ndctl), the driver can operate
only if /sys is mounted in read-write mode.
It has been seen however that sometimes container gets
deployed in such a way that /sys is mounted read-only.
It makes sense to detect writable /sys at start-time
and refuse to run, instead of stumbling to that later in operation.