Persisting driver state and reload state on restart.
Upon restarts, driver has to recognize the provisioned volumes on last run.  For
that driver has to persist its state(provisioned volume information). We have
two options:

1. Per node state: Persist the driver state on each node specific to that node.
Upon driver start, it has to load the persisted information. Whenever a node
registers, Controller/Master has to collect the volume information stored on
that node.

2. Cluster level state: Persist the consolidated cluster level driver state, say
to a configmap on Kubernetes. This is supposed to done at Controller side, but not
clear how to pass node specific information to a node.

This change implements the fist option, i.e, per node state.

Defined a new interface, StateManager, that handles the state persistency
irrespective of either per node or cluster. Also provided an implementation of
StateManager that persists the state to a file.

Made changes to RegistrySrerver, so that it informs the listeners upon
registering of a node controller. The ControllerServer on master, listens on
this and collects the volume information on node, by calling
csi.ControllerServer.ListVolumes.

FIXES #25 - Clean handling of driver restart