docs: reformat + full table of contents
Future diffs will be much shorter when lines are short. Changing a
single word in a line that spans an entire paragraph shows the entire
paragraph as changed. Short lines are also more readable in a shell.

The table of contents was generated by markdown-toc-generate-toc in
Emacs. It includes all headings.