docs: avoid "plugin", improve intro, consistent spelling
PMEM-CSI is not a "Kubernetes plugin". That is something else. Whether
we want to call it "driver" (as in "Kubernetes storage driver") or
"plugin" (as in the CSI spec) is debatable. For our
audience (Kubernetes admins and users) "driver" makes more sense and
it also seemed to be more common already in the documentation, so
"plugin" gets replaced by that or just PMEM-CSI.

Capitalization is PMEM-CSI when using the name. Only when referring to
files or Kubernetes objects, pmem-csi (lower case) is used.