pmem-csi-driver: support correct target directory behavior
Kubernetes implements the target directory behavior for
NodePublishVolume incorrectly: instead of letting the driver create
the directory, it gets created by
kubelet (https://github.com/kubernetes/kubernetes/issues/75535).

A standard-compliant driver has to create and delete the
directory. But as we want to be compatible with Kubernetes, we also
allow the directory to exist already.

csi-sanity from csi-test v2.0.0 will check for the correct behavior.