test: install cluster into sub-directory
This moves all files related to a Clear KVM cluster into a
configurable sub-directory of _work. This makes it possible to keep
two clusters around and paves the way towards setting up the cluster
in a different way (for example, with GoVM).

The troubleshooting section for QEMU gets removed because it will
become obsolete once we use GoVM and it no longer works exactly as
described.