Write termination message to TERMINATION_MESSAGE_PATH
The path is different from default /dev/termination-log
because we need to pass /dev/ explicitly to containers,
and that is known to cause container creation failure
if combined with termination-log located on /dev.
So we relocate termination log to /tmp.
We define env.var that we pass into container so that
container knows where to create the termination message.