remove README-qemu-notes.md and documentation of unified mode
It isn't useful anymore and would have to be updated to fix the
out-dated pmem-CSI spelling. The only useful information seems to be
the information why ndctl must be used after the first boot, which was
moved into a comment in start-stop.make.

The only reference to the file was for unified mode, which is
scheduled to be removed (#256), so that entire section gets removed.