devicemanager/lvm: Ask lvm not to clear newly created volume
On non-alpine images, Due to some race conditions lvcreate fails to clear the
newly created block device as the result, volume creation fails.

https://linux-lvm.redhat.narkive.com/9Pu8Wro2/lvcreate-device-not-cleared-aborting-failed-to-wipe-start-of-new-lv

As we explicitly clearing the device in our code, we need not depend on lvm's
device clearing, hence this change ask's lvcreate not to zero the device.