Dockerfiles: invalidate cache before "swupd update"
Docker caches previous build results and skips commands that it has
executed in the same context before. The result is that the "swupd
update" command may get skipped during a build:

Step 6/33 : RUN swupd update && swupd bundle-add ${NDCTL_BUILD_DEPS} go-basic-dev && rm -rf /var/lib/swupd
 ---> Using cache
 ---> 229de2b1ee2e

That caching defeats the purpose of the command, which is to ensure
that all the latest updates for the base image are getting applied.

On the other hand, not using a cache at all (= "docker build
--no-cache") would slow down developers a lot (> 20min build times).

The solution is to invalidate the cache right before the "swupd
update" command by injecting an unused "CACHEBUST" variable, as
described in http://dev.im-bot.com/docker-select-caching/. The default
is to invalidate the cache once per day. A CI build can invalidate it
for each build.

In the Makefile, the variable is called BUILD_IMAGE_ID because that's
what it is; "busting the cache" is simply what it is currently used
for.