test: expose controller interface on csi.sock only during testing
The controller interface on a node's csi.sock is only useful for
testing, therefore it is better to also only add it when running in
that mode.

kustomize doesn't support patching the same thing (like arguments)
twice with different patches. Therefore the patches which previously
only increased the log level get generalized to "add arguments for
testing", which now also includes setting the new -testEndpoint flag.