test: check behavior of controller after restart
The nodes are supposed to reconnect to the
controller (https://github.com/intel/pmem-CSI/commit/06c0d7ab0645f24c0cbc91d33adeb0547863c4f8).
This test verifies whether they do that within 2 minutes after the
controller is up again. If they don't, it fails like this:

• Failure [67.317 seconds]
sanity
/nvme/gopath/src/github.com/intel/pmem-csi/test/e2e/storage/sanity.go:48
  pmem-csi
  /nvme/gopath/src/github.com/intel/pmem-csi/test/e2e/storage/sanity.go:146
    capacity is restored after controller restart [It]
    /nvme/gopath/src/github.com/intel/pmem-csi/test/e2e/storage/sanity.go:240

    Timed out after 60.000s.
    total capacity after controller restart
    Expected
        <int64>: 28978446336
    to equal
        <int64>: 86935339008

    /nvme/gopath/src/github.com/intel/pmem-csi/test/e2e/storage/sanity.go:254

Note that the controller is in a bad state after this test failure and
the cluster needs to be restarted!