test: fix "make restart"
When Kubernetes failed to come up (for example, because of
https://github.com/clearlinux/distribution/issues/891), the check for
running Kubernetes incorrectly proceeded because it found no pending
nodes. The error about not being able to connect to the API server was
suppressed, too.

Now the logic is inverted: the expected number of nodes must be up.

Checking for pods isn't very useful and gets removed.