test: avoid invalid TEST_LOCAL_REGISTRY
When there is no docker0 interface as in the Jenkins CI, we end up
adding :5000 to the Docker config, which is invalid and causes Docker
to not start. It's okay to not have the correct default here because
it's not used, but the default must be valid. Use localhost as
fallback.