pmem-csi-driver: fix race condition in registry server
When nodes register, there is a race condition around the concurrent
modifications and/or usage of the node info map by the master
controller. All accesses to that field have to be protected by a
mutex.

To enforce this, the RegistryServer implementation gets moved into its
own package with only those methods exported which do proper
locking.