Jenkinsfile: simplify environment, remove cleanup and image version
We run in a virtual machine, so each job has its own Docker instance.
Therefore we don't need to worry about image version conflicts and we
don't need to clean up.

The ${env.<variable name>} syntax inside double-quoted strings is used
to ensure Jenkins directly expands the command, instead of relying on
passing the variables into the shell and expanding there. The
advantage is that the Jenkins log will show the expanded command,
which makes it easier to debug the Jenkins pipeline.