deploy: enable late binding example
For Kubernetes 1.14 (and only there) should we show how to try out the
late binding mode. On 1.13 it doesn't work reliably.

This becomes easier when describing the deployment inside the setup
script (just one place which needs to be modified, not two).

While at it, the .yaml files also get cleaned up (no empty line at the
end, same comment as in other .yaml files).