pmem-csi-driver: clarify and fix callbacks in registry
The mutex that protect RegistryServer data structures must not be held
while invoking callbacks with unknown code. Doing that has been found
to cause deadlocks when that code itself calls the RegistryServer
again.

This change should be enough to prevent the deadlock, but further work
might be needed to make the callbacks resilient against concurrent
invocations.