deploy: increase external-provisioner gRPC timeout
In particular shredding large volumes can take a long time. To avoid
premature timeouts and the ensuing re-transmission + exponential
backoff, we can increase the default timeouts. How to do that depends
on the external-provisioner version.

The 5 minutes value was chosen rather arbitrarily. Testing on real
hardware is needed to check whether that is a good value.