e2e: log failures as soon as they occur
When a Gomega assertion fails, it calls a failure handler. The
Kubernetes wrapper extends the information that is capture, but
otherwise just passes the failure through to Gingko.Fail, which
records the failure for the current test and then aborts the test.

This has two drawbacks:
- when looking at the output of a test with additional cleanup
  actions, it is not obvious why or where the test stopped
  because Ginkgo only logs the failure after it is done
  with the test
- if cleanup actions encounter additional assertion failures (not
  good style because cleaning up then aborts, but may happen),
  those failures are never logged because Ginkgo only remembers
  the initial failure

To solve both, the extended failure handler always logs the failure
before proceeding. This is a more general solution than
logger.go:Failf() because it works for all kinds of failures, not just
those triggered by a call to that function.

This is a modified copy of
k/k/test/e2e/framework/ginkgowrapper/wrapper.go with some code copied
from k/k/test/e2e/framework/log/logger.go. The goal is to submit these
enhancements back to Kubernetes.