test: increase deadline for sanity stress test
The current deadline was to low when testing in the CI. Besides being
slower overall, the formula also didn't take into account that
shredding becomes slower when there are multiple pending shredding
operations.

In order to get a feeling for how slow each individual volume test is,
the duration now gets measured and logged.