CI: push images in non-PR jobs
All post-merge and periodic jobs push the new images if the job was
successful. Then some Intel-internal CI can pull it and promote it to
the official download location on Docker Hub.

Because the name of the images now include the registry, also the test
jobs need to know the registry name.