build: limit Docker context to files that are needed
In our Dockerfile we only need the Go source code, make rules and
configuration files. By explicitly selecting those we can cut down the
amount of data sent to the Docker daemon from 120MB to 56MB.

Another advantage is that changes to other files no longer invalidate
the Docker cache for "ADD ." because the modified files are not
included.