QEMU: enhance documentation and check for /dev/kvm
Not having KVM enabled in a GCE machine is a real issue that a user
ran into. Error handling in this case is poor, GoVM doesn't print any
error and all that the user gets is an error about timing out waiting
for ssh.

The updated README.md explains the real requirements better and warns
about this pitfall. The script itself checks very early for /dev/kvm
and bails out with an error message when it is missing.