Use klog for logging
Driver logs are not logging the real filename from where the message generated.
Because Kubernetes glog is just a wrapper around klog and klog implements the
logic to fetch the file name from it's caller. As in our case all log messages
are logged via glog, hence fails to find the real filename.

This change fixes this by using directly klog's logging API.