Improve PublishVolume,StageVolume to meet CSI spec more closely
If a volume is already mounted, we must check are the
parameteres compatible, and return OK if they are.
For having something to compare, we have to store params on Publish.
Use sorted and converted to single string mount options in compare.
Start with mount options given in request, instead of empty,
to take given options into use if present.
fsType is not stored in settings, but detected from device.