test: Fedora support
This generalizes the concept of choosing a distro (via the new
TEST_DISTRO) and its version (with TEST_DISTRO_VERSION instead
of TEST_CLEAR_LINUX_VERSION, which is still supported for people
following older instructions).

The first new distro is Fedora. We use the upstream Docker and
Kubernetes, which will make it possible to also install specific
Kubernetes versions.