CI: test older Kubernetes with Fedora
Using an older Clear Linux release to get also an older Kubernetes was
a bit hacky and not realistic, because no-one is going to use such an
installation. Now that we have Fedora support and can choose the
Kubernetes version that gets installed on it, we no longer need that
hack.

The pull request testing continues to use a recent Clear Linux.