CI: install kubectl on host
We run the tests inside the build container, which normally doesn't
have kubectl installed. It shouldn't need that either because the
e2e.test binary connects directly to the running cluster. But
currently the volumemode tests depend on kubectl
(https://github.com/kubernetes/kubernetes/issues/83913), so we have to
provide it for now.

The quickest way to "install" kubectl is to copy it from the cluster.