test: lock down version of Kubernetes control plane
The binaries on the host were already locked on Fedora via the
versions that yum was asked to install, but kubeadm (against the
expectation) then still ended picking up the latest control plane
images as long as they had the same <major>.<minor> versions.

For full reproducibility, we also want the control plane to be exactly
from the same release as the host binaries.