CI: enable JUnit output, enhance and fix TEST_E2E make parameters
We have to run with a custom reporter in order to write JUnit
files. This will be useful for reporting in the CI what tests we ran.
While at it, add TEST_E2E_FOCUS and fix TEST_E2E_SKIP (variable was
spelled wrong).

The junit Jenkins plugin interprets a dot as separator between test
groups. We want all runs to be shown at the same level, therefore we
have to avoid the dot and replace it with an underscore.