CI: no logging by default
The recently added increased logging made the resulting
log files rather large, which is a problem when downloading
them and exhausted available space in the Jenkins master.

Now that extra logging is turned off by default respectively
done less frequently (top and df run once a minute). To enable
logging, edit the pipeline in a PR or in the Jenkins UI.