NodeStageVolume: return OK if same device already mounted
CSI spec requires StageVolume to be idempotent, so we have
to check for same volume mount and capability match and
return OK if already mounted.
Move existing FsType and mount checking out of ProvisionVolume
as this is only relevant on path starting from NodeStageVolume
(as another path uses freshly-created Volume which can not be mounted).
Pass codes.error up from ProvisionVolume so that better suited error
code can be returned by NodeStageVolume.