images: fix content of /usr/local
Because both PMEM binaries and the Go toolchain were installed into
/go and the runtime image contained all of /go/bin, we ended up also
copying the Go toolchain.

License files for everything that goes into our /usr/local are
important. 3-clause BSD requires that the copright notice is
reproduced in "other materials provided with the distribution", which
we comply with by copying the license file and its copyright remark
into the image.

hack/copy-modules-license was originally developed for
https://github.com/intel/intel-device-plugins-for-kubernetes/pull/232
and adapted to PMEM-CSI here (multiple packages, no dependency on "go
mod", etc.)