vendor-bom.csv: track actual licenses
We used to base the analysis on the modules imported by dep. With the
switch to go mod vendoring we can no longer do that and it also wasn't
quite complete: some of the imported modules embed code from elsewhere
which sometimes has other licenses.

Therefore we now look for license files and list anything that has a
separate license.