test: run Go tests inside driver container
This has multiple advantages:
- we test in exactly the environment in which the code will run
- it's faster than bringing up a separate VM
- no need to remember a separate make target to run all tests

Speed is important both for CI and developers. Running the tests in an
existing cluster just takes 9 seconds and can be done using the same
command as running other tests individually:

   make test_e2e TEST_E2E_FOCUS=gotests.*pmem-device-manager

Because it runs so quickly, it's okay to simply repeat this testing in
each CI step which runs E2E tests.