build: warn about Go version
A hard check is too strict (Go 1.14 might work) whereas a >= version
check can fail (1.13 vendored differently than 1.12 and thus couldn't
be instead of that one). What we do instead is warn when the
major+minor parts of the version number differ, without aborting the
actual build.

The "hack" directory should be small enough that we can make all of it
available for Docker builds. That way we don't need to whitelist
individual files.