deploy: support kustomize for end-users
"kubectl kustomize" cannot be called for all bases in deploy/kustomize
because it has no `--load_restrictor none` and the security policy prevents
reading some patch files:

  $ kubectl kustomize deploy/kustomize/kubernetes-1.14-lvm-testing/
  Error: AccumulateTarget: AccumulateTarget: AccumulateTarget: security; file '../../patches/controller-role-patch.yaml' is not in or below '/nvme/gopath/src/github.com/intel/pmem-csi/deploy/kustomize/kubernetes-1.14/rbac'

We can solve this for users by creating "normal" bases where all
patches and resources have already been pulled into a single
pmem-csi.yaml. All we need to do is to put them into a separate
directory and add a kustomize.yaml.

For the sake of preserving current behavior we store those generated
files also under their old names. Symlinks are not used because that
does not work on Windows or when pulling the files via http.