test: double size of PMEM, use one half for LVM
This verifies that kustomize works for our base deployments. It also
adds some limited testing of the -usefordax and -useforsector
parameters because the now unused space will be used for runnning the
device manager Go test. If all space was used for LVM, those tests
would fail when using direct mode.