ndctl: Consider only non-zero sized namespaces while retrieving namespaces
While deleting a namespace ndctl might still keeep the namespace inactive
and with size zero. And also there might be possible that the unsuccessful
deletes via Region.DestroyNamespace() might left the namespace disabled but not
deleted, so we should not ignore them too.

This change, considers the size of the found namespace and considers only if
it's non-zero.