test: check that volumes really support DAX
We could check mount options, but that wouldn't work for raw block
volumes. A better approach is to do exactly what an application would
do, which is calling mmap(MAP_SYNC). A new Go program is used for that
because then we can compile on the host and run on the target. This
new pmem-dax-check might also be useful as stand-alone program in
_output, but for now it just gets build on demand when running tests.

The check is implemented as a test suite that applies the same test to
several differen test patterns (different file systems,
PVC/inline, raw blocks). Raw blocks need special treatment because
dax only works after mounting.