CI: run under normal user inside Docker
The recently added vendor check rewrites "vendor" inside Docker, which
turned out to break the job which in later steps need to work with the
repo because those files then became owned by root.

We could run "sudo chown" after each "docker run" invocation, but it
is safer and faster to not run as root inside Docker in the first
place.

By setting GOPATH such that it ends up on the host system we also
should get better performance because the cache can be reused between
different steps. Previously, /.cache was used by Go, which is
a local directory inside the container which does not persist
across "docker run" invocations.

Installing ssh gets moved to the Dockerfile to avoid the need for
"sudo" inside the build container and to speed up CI testing (no need
to install it over and over again). With current Clear Linux we can
use "openssh-client" which only adds 4MB, so it's okay to do this
unconditionally even if it might no be needed (for example, by
developers).