test: switch to cri-o on recent Clear Linux
This is using the (at this time) latest Clear Linux, which brings up a
1.16 Kubernetes cluster. A test with Docker was not stable (containerd
went down during testing on one node, which then caused Docker and
kubelet to also fail) and upstream is mostly testing with cri-o and it
should be more stable now, so let's also use that.

A recent flannel and systemd as cgroup driver were needed to bring
up a functional cluster.

The defaults in test-config.sh and comments get adapted accordingly
and cleaned up a bit.