CI: run without sudo
The last remaining reason for requiring sudo and --privileged was that
the scripts themselve set up a tmpfs for etcd. By moving that step out
into the build host we can remove the need for sudo inside the
container.

The etcd volume gets used repeatedly, therefore "mkfs.ext4" must be
forced to ignore an already existing filesystem.