CI: reuse a running cluster when possible
Instead of always shutting down the cluster after a test, we keep it
running. Then the next test can check whether it can reuse the cluster
and only shuts it down when it can't.

We can redeploy PMEM-CSI with different device and deployment modes,
so those parameters can be ignored. What matters is the distro and
Kubernetes release.

The advantage is that we save around 5 minutes (for Fedora, the more
common case) in roughly half of the steps.