CI: skip slow tests in PR jobs
We have some tests that are marked as "slow", for example those that
always wait 5 minutes for something to *not* happen because PV
creation is expected to not work.

Skipping those tests is an easy way to speed up the pre-merge
testing. We loose of course some coverage, but that's what we have
full testing of the "devel" branch for before moving the change into
"master" or doing a release.