remove entry point for PMEM-CSI image
It's very annoying that the pmem-csi-driver is marked as entry point
for the pmem-csi-driver image when using it as the container for
running tools that aren't installed on a host, like ndctl or LVM. For
example:

   $ docker run --privileged  --rm intel/pmem-csi-driver:v0.5.0 -- ndctl list -Nu
   failed to initialize driver: Invalid driver mode: unified

It's better to not specify an entry point and instead select the
command in the deployment. This is then also consistent for all three
commands that we run from the image.

Fixes: #377