CSI driver: unify and enhance volume parameter handling
Validation of user input was missing, with the result that misspelled
keys were silently ignored.

`persistencyModel: none` was supported, but not documented. `none` is
a misleading name, so it gets renamed to `normal` (a breaking
change!).

Support for units in the ephemeral inline volume size was undocumented
and (against expectations) was implemented such that 1M expanded to
1000 * 1000 * 1000 (decimal unit). In this context binary units make
more sense, so those are used now (another breaking change).

There was support for `persistencyModel: ephemeral` in the master
controller, but that value is invalid there, so now this triggers an
error and all code around ephemeral volumes in the master controller
could be removed.

Fixes: #501