imagefile: remove second MBR
The first MBR may be useful when working with the image file on the
host, but the second one has no real purpose. It just causes more work
for the kernel and the kernel developers were even contemplating to
remove or change partitioning support. Even though it looks like it
will be kept (https://lkml.org/lkml/2020/1/9/904), we better shouldn't
depend on it unnecessarily.