support Kata Containers
A persistent or ephemeral volume can either be prepared for usage by
DAX-enabled applications that don't run under Kata Containers (the
default) or for DAX-enabled applications that run under Kata
Containers.

In both cases the volume can be used with and with Kata Containers,
it's just that DAX only works either inside or ourside of Kata
Containers.

The Kata Container runtime must be able to access the image file while
it is still mounted, therefore we cannot use something inside the
target dir as mount point, because then the image file is shadowed by
the mounted filesystem.

We already have a local state dir for .json files. Putting something
else inside it might confuse the state code, so instead we create a
second directory with ".mount" appended to the directory name and use
that for mount points.

We also have to enable bi-directional mount propagation for it because
otherwise the mounted fs with the image file is still only visible
inside the container).