docs: describe latest release and how to create a new one
It's worthwhile to state right at the beginning of our main landing
page (= README.md) what we currently support and how we handle tags
because users need to know that when picking something to install.

Doing a release involves certain steps. We had documentation for most
of that, but not a single simple check list.